Update MatterPostAttributeChangeCallback() function

+++ examples/all-clusters-app/ameba/main/CHIPDeviceManager.cpp
@@ -87,7 +87,7 @@ exit:
 } // namespace DeviceManager
 } // namespace chip
 
-void MatterPostAttributeChangeCallback(const chip::app::ConcreteAttributePath & path, uint8_t type, uint16_t size, uint8_t * value)
+void MatterPostAttributeChangeCallback(const chip::app::ConcreteAttributePath & path, uint8_t type, uint16_t size, uint16_t * value)
 {
     deviceCallback.PostAttributeChangeCallback(path.mEndpointId, path.mClusterId, path.mAttributeId, type, size, value);
 }
